backend/requirements.txt
certifi==2019.3.9
cffi==1.13.1
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pycparser==2.19
pyOpenSSL==19.0.0
query-string==2019.4.13
requests==2.22.0
six==1.12.0
urllib3==1.25.6
Werkzeug==0.16.0
AuthLib==0.12.1
python-dotenv==0.10.3
numpy>=1.13.3
python-dateutil>=2.6.1
pytz>=2017.2
pandas==0.25.3
xlrd==1.2.0
flask-cors==3.0.8
pyyaml==5.1.2 
Flask-SQLAlchemy==2.4.1 
SQLAlchemy==1.3.11
psycopg2==2.8.4
Flask-Script==2.0.6
Flask-Migrate==2.5.2 
Mako==1.1.0 
alembic==1.3.1 
python-editor==1.0.4
psycopg2-binary==2.8.4
pyjwt==1.7.1
gspread==3.1.0
aniso8601==8.0.0 
flask-restful==0.3.7
scipy==1.4.1
requests-cache==0.5.2
multitasking==0.0.9 
lxml==4.4.2
html5lib==1.0.1
webencodings==0.5.1
beautifulsoup4==4.8.2
bs4==0.0.1
soupsieve==1.9.5
prometheus-client==0.7.1
prometheus-flask-exporter==0.11.0
py-healthcheck==1.9.0